refactor(lens): migrate all legacy Executor tool methods to per-domain free functions (closes #1655)
Consolidates PRs 1-9 of epic #1655 into a single commit. Migrates all
44 remaining `_tool_*` methods off `class Executor` and onto top-level
`def impl(ctx, ...)` functions in the matching lens/<domain>.py files
— the same pattern new tools (list_my_runs, actor mutators) already
use.

## What changed

**42 tools migrated across 7 domain files** (policies.py already
migrated via #1657 on main):
- discovery: get_discovery_summary (1)
- marketplace: list_installed_packs, browse_marketplace, get_pack_details (3)
- workflows: list_workflows, get_blocked_workflows, get_workflow_details,
  list_runs, get_run, list_run_events (6)
- workspace: list_agent_identities, get_agent_identity_count,
  list_integrations, get_integration_status, list_members, get_member,
  list_projects, get_project (8)
- ops: list_pending_approvals, get_approval, get_audit_events,
  search_audit_log, list_alerts, get_alert (6)
- governance: get_compliance_status, get_governance_kpis,
  get_framework_coverage, get_governance_narrative,
  get_recent_governance_events (5)
- guard_core: get_spend_summary, get_recent_events, get_sessions,
  get_event_count, search_memory, search_sessions, get_team_memory_feed,
  get_session_reports_feed, search_knowledge, get_guard_config,
  get_budgets, get_correlated_events, get_savings_summary (13)

**Cleanup:**
- `_impl()` and `_run()` shims deleted from `_shared.py`
- `class Executor` shrunk from 1258 LOC to a 15-line context holder
  (still constructed by chat.py to carry db/workspace_id/agent_identity_id)
- `MIN_SIMILARITY_SCORE` class constant moved to module-level in guard_core.py
- `opener.py` rewired to call `get_governance_kpis(ctx)` directly instead
  of `executor._tool_get_governance_kpis()`
- Dead `_impl`/`_run` imports stripped from every registration file

**Test updates:**
- `test_executor_guard_check.py` — mocks now patch `tool.impl` on the
  ToolDef directly (not the module attribute) since ToolDef captures
  the function object at registration time
- `test_lens_events_tool.py`, `test_lens_caching_and_approvals.py`,
  `test_agent_identity_tools.py`, `test_run_tools.py`,
  `test_workflow_tools.py` — auto-rewritten from `executor._tool_X(...)`
  to `X(executor, ...)` (Executor still has .workspace_id, so it works
  as a ctx). Where needed, tests now patch `app.core.database.SessionLocal`
  to inject a fake db.

## Why one PR instead of nine

Original plan was 8 domain PRs + 1 cleanup. First-shipped PR (#1657
policies) merged cleanly. Remaining branches conflicted with each
other on executor.py delete-boundaries when trying to consolidate.
Simpler and safer to ship the rest as one atomic PR — the changes are
purely mechanical and all-or-nothing at the file level anyway.

## Verified

- All previously-failing tests from the epic pass (219 tests in the
  affected test files)
- Full suite: 4139 pass on this branch. The 107 remaining failures
  are identical to origin/main baseline (schema drift on local dev DB,
  missing prometheus_client, and pre-existing test breakage) — zero
  new failures introduced by this refactor.
- `class Executor` now has ZERO `_tool_*` methods
- No `impl=_impl("...")` shim references remain
