+++ apps/api/app/tools/registrations/lens/workspace.py
@@ -10,8 +10,6 @@ from __future__ import annotations
 
 from app.tools.types import ToolDef
 from app.tools.registrations.lens._shared import (
-    _impl,
-    _run,
     _actor_impl,
     _window_start,
     _LIMIT,
@@ -103,13 +101,186 @@ def get_workspace_kpis(ctx, time_window: str = "last_24h"):
         db.close()
 
 
+# ── Migrated from Executor (epic #1655 PR 5/9) ─────────────────────────
+def list_agent_identities(ctx, status: str = "active", limit: int = 20):
+    """List agent identities in this workspace."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.modules.agent_identity.models import AgentIdentity
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        q = db.query(AgentIdentity).filter(AgentIdentity.workspace_id == ws_uuid)
+        status = (status or "active").lower()
+        if status != "all":
+            q = q.filter(AgentIdentity.lifecycle_state == status)
+        rows = q.order_by(AgentIdentity.created_at.desc()).limit(min(limit, 100)).all()
+        return [
+            {
+                "id": r.id,
+                "name": r.name,
+                "token_prefix": r.token_prefix,
+                "lifecycle_state": r.lifecycle_state,
+                "risk_tier": r.risk_tier,
+                "source": r.source,
+                "platform_of_origin": r.platform_of_origin,
+                "created_at": r.created_at.isoformat() if r.created_at else None,
+                "deactivated_at": r.deactivated_at.isoformat() if r.deactivated_at else None,
+                "last_used_at": r.last_used_at.isoformat() if r.last_used_at else None,
+                "expires_at": r.expires_at.isoformat() if r.expires_at else None,
+            }
+            for r in rows
+        ]
+    finally:
+        db.close()
+
+
+def get_agent_identity_count(ctx, status: str = "active"):
+    """Exact COUNT of agent identities matching lifecycle_state."""
+    import uuid as _uuid
+    from sqlalchemy import func as sa_func
+    from app.core.database import SessionLocal
+    from app.modules.agent_identity.models import AgentIdentity
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        q = db.query(sa_func.count(AgentIdentity.id)).filter(
+            AgentIdentity.workspace_id == ws_uuid
+        )
+        status = (status or "active").lower()
+        if status != "all":
+            q = q.filter(AgentIdentity.lifecycle_state == status)
+        return {"count": int(q.scalar() or 0), "status": status}
+    finally:
+        db.close()
+
+
+def list_integrations(ctx):
+    """All configured integrations for the workspace."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.models.integration import Integration
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        rows = db.query(Integration).filter(Integration.workspace_id == ws_uuid).all()
+        return [{"id": str(r.id), "service": r.service, "auth_method": r.auth_method,
+                 "handle": r.handle, "scopes": list(r.scopes) if r.scopes else [],
+                 "last_used_at": r.last_used_at.isoformat() if r.last_used_at else None,
+                 "created_at": r.created_at.isoformat() if r.created_at else None}
+                for r in rows]
+    finally:
+        db.close()
+
+
+def get_integration_status(ctx, service: str):
+    """Status + metadata for one integration by service slug."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.models.integration import Integration
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        r = db.query(Integration).filter(
+            Integration.workspace_id == ws_uuid, Integration.service == service.lower()
+        ).first()
+        if not r:
+            return {"service": service, "configured": False}
+        return {"service": r.service, "configured": True, "auth_method": r.auth_method,
+                "handle": r.handle, "scopes": list(r.scopes) if r.scopes else [],
+                "last_used_at": r.last_used_at.isoformat() if r.last_used_at else None}
+    finally:
+        db.close()
+
+
+def list_members(ctx, role: str | None = None, limit: int = 50):
+    """List workspace members. Optional role filter."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.models.workspace_user import WorkspaceUser
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        q = db.query(WorkspaceUser).filter(WorkspaceUser.workspace_id == ws_uuid)
+        if role:
+            q = q.filter(WorkspaceUser.role == role.lower())
+        rows = q.order_by(WorkspaceUser.joined_at.desc()).limit(min(limit, 200)).all()
+        return [{"clerk_user_id": r.clerk_user_id, "role": r.role, "invited_by": r.invited_by,
+                 "joined_at": r.joined_at.isoformat() if r.joined_at else None}
+                for r in rows]
+    finally:
+        db.close()
+
+
+def get_member(ctx, clerk_user_id: str):
+    """One member by clerk_user_id."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.models.workspace_user import WorkspaceUser
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        r = db.query(WorkspaceUser).filter(
+            WorkspaceUser.workspace_id == ws_uuid, WorkspaceUser.clerk_user_id == clerk_user_id
+        ).first()
+        if not r:
+            return {"error": "Member not found"}
+        return {"clerk_user_id": r.clerk_user_id, "role": r.role,
+                "role_id": str(r.role_id) if r.role_id else None,
+                "invited_by": r.invited_by,
+                "joined_at": r.joined_at.isoformat() if r.joined_at else None}
+    finally:
+        db.close()
+
+
+def list_projects(ctx, limit: int = 50):
+    """List projects in the workspace."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.models.project import Project
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        rows = (db.query(Project).filter(Project.workspace_id == ws_uuid)
+                .order_by(Project.created_at.desc()).limit(min(limit, 200)).all())
+        return [{"id": str(r.id), "name": r.name, "slug": r.slug,
+                 "project_type": r.project_type,
+                 "created_at": r.created_at.isoformat() if r.created_at else None}
+                for r in rows]
+    finally:
+        db.close()
+
+
+def get_project(ctx, id_or_slug: str):
+    """One project by UUID or slug."""
+    import uuid as _uuid
+    from app.core.database import SessionLocal
+    from app.models.project import Project
+    db = SessionLocal()
+    try:
+        ws_uuid = _uuid.UUID(ctx.workspace_id)
+        q = db.query(Project).filter(Project.workspace_id == ws_uuid)
+        try:
+            r = q.filter(Project.id == _uuid.UUID(id_or_slug)).first()
+        except ValueError:
+            r = q.filter(Project.slug == id_or_slug).first()
+        if not r:
+            return {"error": "Project not found"}
+        return {"id": str(r.id), "name": r.name, "slug": r.slug,
+                "project_type": r.project_type,
+                "security_finding_id": r.security_finding_id,
+                "created_at": r.created_at.isoformat() if r.created_at else None}
+    finally:
+        db.close()
+
+
 # ── ToolDef list ───────────────────────────────────────────────────────
 TOOLS: list[ToolDef] = [
     ToolDef(
         name="list_integrations",
         description="All integrations (Slack, GitHub, Okta, Vercel, ...) configured for this workspace.",
         input_schema={"type": "object", "properties": {}, "required": []},
-        impl=_impl("list_integrations"),
+        impl=list_integrations,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -121,7 +292,7 @@ TOOLS: list[ToolDef] = [
             "properties": {"service": {"type": "string", "description": "e.g. github, slack, okta, vercel"}},
             "required": ["service"],
         },
-        impl=_impl("get_integration_status"),
+        impl=get_integration_status,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -136,7 +307,7 @@ TOOLS: list[ToolDef] = [
             },
             "required": [],
         },
-        impl=_impl("list_members"),
+        impl=list_members,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -148,7 +319,7 @@ TOOLS: list[ToolDef] = [
             "properties": {"clerk_user_id": {"type": "string"}},
             "required": ["clerk_user_id"],
         },
-        impl=_impl("get_member"),
+        impl=get_member,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -160,7 +331,7 @@ TOOLS: list[ToolDef] = [
             "properties": {"limit": _LIMIT},
             "required": [],
         },
-        impl=_impl("list_projects"),
+        impl=list_projects,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -172,7 +343,7 @@ TOOLS: list[ToolDef] = [
             "properties": {"id_or_slug": {"type": "string"}},
             "required": ["id_or_slug"],
         },
-        impl=_impl("get_project"),
+        impl=get_project,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -195,7 +366,7 @@ TOOLS: list[ToolDef] = [
             },
             "required": [],
         },
-        impl=_impl("list_agent_identities"),
+        impl=list_agent_identities,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
@@ -215,7 +386,7 @@ TOOLS: list[ToolDef] = [
             },
             "required": [],
         },
-        impl=_impl("get_agent_identity_count"),
+        impl=get_agent_identity_count,
         annotations=_READ_ONLY,
         tags=_LENS_TAGS,
     ),
