+++ apps/api/tests/test_lens_caching_and_approvals.py
@@ -1,10 +1,32 @@
-"""Tests for Lens caching + approvals-tool broadening (2026-08-28 follow-up to #1347)."""
+"""Tests for Lens caching + approvals-tool broadening (2026-08-28 follow-up to #1347).
+
+Post epic #1655: tools migrated from Executor to top-level free functions
+in lens/ops.py. Tests patch app.core.database.SessionLocal to inject a
+mock DB, since the tools now open their own session.
+"""
 from __future__ import annotations
 
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
+
+from app.tools.registrations.lens.ops import list_pending_approvals
 
 # Import llm_client first so its re-exports resolve before we touch adapters.
 import app.runtime.llm_client  # noqa: F401
+
+
+class _Ctx:
+    workspace_id = "00000000-0000-0000-0000-000000000001"
+
+
+def _fake_db_with_query():
+    fake_query = MagicMock()
+    fake_query.filter.return_value = fake_query
+    fake_query.order_by.return_value = fake_query
+    fake_query.limit.return_value = fake_query
+    fake_query.all.return_value = []
+    fake_db = MagicMock()
+    fake_db.query.return_value = fake_query
+    return fake_db, fake_query
 
 
 def test_adapter_default_enables_cache_system():
@@ -24,21 +46,9 @@ def test_adapter_default_enables_cache_system():
 
 def test_list_pending_approvals_supports_since_today():
     """since='today' should filter created_at >= UTC start-of-day."""
-    from app.modules.glens.executor import Executor
-
-    executor = Executor.__new__(Executor)
-    executor.workspace_id = "00000000-0000-0000-0000-000000000001"
-
-    fake_query = MagicMock()
-    fake_query.filter.return_value = fake_query
-    fake_query.order_by.return_value = fake_query
-    fake_query.limit.return_value = fake_query
-    fake_query.all.return_value = []
-
-    executor.db = MagicMock()
-    executor.db.query.return_value = fake_query
-
-    result = executor._tool_list_pending_approvals(status="all", since="today", limit=10)
+    fake_db, fake_query = _fake_db_with_query()
+    with patch("app.core.database.SessionLocal", return_value=fake_db):
+        result = list_pending_approvals(_Ctx(), status="all", since="today", limit=10)
     assert result == []
     # workspace_id + since (status skipped for "all") = 2 filter calls
     assert fake_query.filter.call_count == 2
@@ -46,37 +56,17 @@ def test_list_pending_approvals_supports_since_today():
 
 def test_list_pending_approvals_supports_iso_date():
     """since='2026-08-28' parses cleanly and applies a filter."""
-    from app.modules.glens.executor import Executor
-
-    executor = Executor.__new__(Executor)
-    executor.workspace_id = "00000000-0000-0000-0000-000000000001"
-    fake_query = MagicMock()
-    fake_query.filter.return_value = fake_query
-    fake_query.order_by.return_value = fake_query
-    fake_query.limit.return_value = fake_query
-    fake_query.all.return_value = []
-    executor.db = MagicMock()
-    executor.db.query.return_value = fake_query
-
-    executor._tool_list_pending_approvals(status="pending", since="2026-08-28")
+    fake_db, fake_query = _fake_db_with_query()
+    with patch("app.core.database.SessionLocal", return_value=fake_db):
+        list_pending_approvals(_Ctx(), status="pending", since="2026-08-28")
     # workspace_id + status + since = 3 filter calls
     assert fake_query.filter.call_count == 3
 
 
 def test_list_pending_approvals_invalid_since_is_ignored():
     """Malformed since string should not crash — silently skip the filter."""
-    from app.modules.glens.executor import Executor
-
-    executor = Executor.__new__(Executor)
-    executor.workspace_id = "00000000-0000-0000-0000-000000000001"
-    fake_query = MagicMock()
-    fake_query.filter.return_value = fake_query
-    fake_query.order_by.return_value = fake_query
-    fake_query.limit.return_value = fake_query
-    fake_query.all.return_value = []
-    executor.db = MagicMock()
-    executor.db.query.return_value = fake_query
-
-    executor._tool_list_pending_approvals(status="pending", since="garbage")
+    fake_db, fake_query = _fake_db_with_query()
+    with patch("app.core.database.SessionLocal", return_value=fake_db):
+        list_pending_approvals(_Ctx(), status="pending", since="garbage")
     # workspace_id + status only; since dropped silently
     assert fake_query.filter.call_count == 2
